Дорабатывает маршрутизацию страницы отзывов и компонент логотипа
