+++ project/src/pages/my-list-screen/my-list-screen.tsx
@@ -37,11 +37,7 @@ function MyListScreen(): JSX.Element {
 
       <footer className="page-footer">
         <div className="logo">
-          <a href="main.html" className="logo__link logo__link--light">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </a>
+          <Logo light/>
         </div>
 
         <div className="copyright">
